backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
google-genai
python-dotenv